requirements.txt
uvicorn[standard]>=0.30.0
python-dotenv